Improve normalization of test and reference results.

+++ test_queries.py
@@ -59,13 +59,14 @@ def test_query(query_id, pytestconfig, asterixdb):
     # Convert result
     df = pd.DataFrame \
         .from_dict(result) \
-        .astype({'x': np.float64, 'y': np.int32})
+        .astype({'x': float, 'y': int})
     logging.info(df)
 
     # Normalize query result
     df = df[df.y > 0]
     df = df[['x', 'y']]
-    df.x = df.x.round(6)
+    df.x = df.x.astype(float).round(6)
+    df.y = df.y.astype(int)
     df.reset_index(drop=True, inplace=True)
 
     # Freeze reference result
@@ -84,7 +85,8 @@ def test_query(query_id, pytestconfig, asterixdb):
     # Normalize reference and query result
     df_ref = df_ref[df_ref.y > 0]
     df_ref = df_ref[['x', 'y']]
-    df_ref.x = df_ref.x.round(6)
+    df_ref.x = df_ref.x.astype(float).round(6)
+    df_ref.y = df_ref.y.astype(int)
     df_ref.reset_index(drop=True, inplace=True)
 
     # Assert correct result
